requirements.txt
streamlit>=1.36
gspread>=6.0.0
oauth2client>=4.1.3
pandas>=2.0.0
pytz>=2023.3
Pillow>=10.0.0
qrcode[pil]>=7.4.2
